requirements.txt
anyio==3.3.4
asgiref==3.4.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
graphql-core==3.1.6
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
strawberry-graphql==0.87.2
twilio==7.3.0
typing-extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
